Finish-README
Script needs some serious debugging work! This is a bit of a blind  project one of those failure is not an option jobs! This is going to take you years and deep down I have the feeling you are just going to fail again. The people who are success in this belong to exclusive clique! But this is what you enjoy doing! So the project will go on for as long as it does until you know longer have access to the Internet. I wish I could fake it I really do now. looking back you where stupid not to take any credit and just go along with things. Used just like a piece of trash. anyway the project is the main thing give it the best shot that is all you can do!